+++ packages/openapi-ts-tests/main/test/plugins.test.ts
@@ -363,6 +363,15 @@ for (const version of versions) {
         description:
           'generate Fetch API client with TanStack Vue Query plugin with custom names',
       },
+      // TODO: add Pinia Colada snapshots
+      // {
+      //   config: createConfig({
+      //     output: 'fetch',
+      //     plugins: ['@pinia/colada', '@hey-api/client-fetch'],
+      //   }),
+      //   description:
+      //     'generate Fetch API client with Pinia Colada plugin',
+      // },
       {
         config: createConfig({
           output: 'default',